made simple terrain using Simplex noise
made a simple terrain generator using simplex noise. For some reason it is being drawn underneath but not on top. (so to see the terrain well go under and look up) that should be an easy thing to fix
The terrain can be scaled. So although it may look very dense with the bumps, if it is scaled up say 1000x, it will look much better.